refactor: terminal 로직 hook으로 분리
